project_sanjaya/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
streamlit
streamlit-folium
pyngrok
PyJWT
bcrypt
python-dotenv
requests
apscheduler
folium
